main script of DCC
Inspired by easy_example. 
Includes DBSCAN and HDBSCAN as clustering algorithms which don't need the prior knowledge about the number of clusters, to compare to.
Includes also a loop which follow the intermediate results of the algorithm 
(looking over the Z and U by using tSNE plot and the above mentioned clustering methods on them).